+++ packages/blocks/src/page-block/edgeless/utils/consts.ts
@@ -6,6 +6,9 @@ export const DEFAULT_NOTE_HEIGHT = 72;
 export const DEFAULT_NOTE_OFFSET_X = 30;
 export const DEFAULT_NOTE_OFFSET_Y = 40;
 
+export const DEFAULT_NOTE_CHILD_FLAVOUR = 'affine:paragraph';
+export const DEFAULT_NOTE_CHILD_TYPE = 'text';
+
 export const FIT_TO_SCREEN_PADDING = 100;
 
 export const ATTACHED_DISTANCE = 20;
